clean load functions
