--- lxl-web/src/routes/(app)/[[lang=lang]]/[fnurgel=fnurgel]/utils.ts
@@ -1,27 +1,0 @@
-import { pushState } from '$app/navigation';
-import isFnurgel from '$lib/utils/isFnurgel';
-
-export function getHoldingsLink(url: URL, value: string) {
-	const newSearchParams = new URLSearchParams([...Array.from(url.searchParams.entries())]);
-	newSearchParams.set('holdings', value);
-	return `${url.origin}${url.pathname}?${newSearchParams.toString()}`;
-}
-
-export function handleClickHoldings(
-	event: MouseEvent & { currentTarget: HTMLAnchorElement },
-	state: object,
-	id: string
-) {
-	event.preventDefault();
-	pushState(event.currentTarget.href, { ...state, holdings: id });
-}
-
-export function getSelectedHolding(value: string, instanceIdsByType: { [key: string]: string[] }) {
-	if (isFnurgel(value)) {
-		return value;
-	}
-	if (instanceIdsByType[value]?.[0]) {
-		return instanceIdsByType[value][0]; // get first instance if showing holdings by type
-	}
-	return undefined;
-}
