+++ lxl-web/src/routes/api/[[lang=lang]]/[fnurgel=fnurgel]/+server.ts
@@ -1,6 +1,7 @@
 import { env } from '$env/dynamic/private';
 import { json } from '@sveltejs/kit';
-import { DisplayUtil, type FramedData, LensType } from '$lib/utils/xl';
+import { DisplayUtil } from '$lib/utils/xl';
+import { type FramedData, LensType } from '$lib/types/xl.types.js';
 import { getSupportedLocale } from '$lib/i18n/locales';
 
 export async function GET({ params, locals }) {
